check_package.py
import pyneo4j_ogm

print("Available in pyneo4j_ogm:")
for attr in dir(pyneo4j_ogm):
    if not attr.startswith('_'):
        print(f"  {attr}")

try:
    from pyneo4j_ogm import Model
    print("✓ Model imported successfully")
except ImportError as e:
    print(f"✗ Model import failed: {e}")

try:
    from pyneo4j_ogm import Property
    print("✓ Property imported successfully")
except ImportError as e:
    print(f"✗ Property import failed: {e}")

try:
    from pyneo4j_ogm import RelationshipTo
    print("✓ RelationshipTo imported successfully")
except ImportError as e:
    print(f"✗ RelationshipTo import failed: {e}")

try:
    from pyneo4j_ogm import Database
    print("✓ Database imported successfully")
except ImportError as e:
    print(f"✗ Database import failed: {e}")
