requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
neomodel==5.4.1
python-dotenv==1.1.1
pydantic==2.11.7
neo4j==5.28.1
